pivotal/mqtt-sub/requirements.txt
pip==20.2.1
paho-mqtt==1.5.0
redis==3.5.2
